Prototype of first JIT instance.
The JITInstance class deals with compilation by creating a queue of tasks before the render() function is called, which completes the tasks and returns the static 'routes'. Pages are hot-reloaded via keeping track of components and templates used.  On a comp/template change, only mentioned pages are used.

Example:

components/
  example.js
pages/
  index.md
  mentionsExample.md

If example.js was changed, only mentionsExample.md would be compiled.